Improve TaskLabel components structure, remove console.logs
