+++ src/components/TaskLabel/DatePicker.tsx
@@ -2,7 +2,6 @@ import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import { Popover, PopoverAnchor, PopoverContent, PopoverTrigger } from "../ui_elements/popover";
 import { format, isValid, parse } from "date-fns";
 import { Input } from "../ui_elements/input";
-import { Button } from "../ui/Button";
 import { CalendarIcon } from "lucide-react";
 import { Calendar } from "../ui_elements/calendar";
