+++ src/components/TaskLabel/TodoTaskLabel.tsx
@@ -1,7 +1,5 @@
-import { useDraggable } from "@dnd-kit/core";
-import { EMCategory, Task } from "../../types";
-import {CSS} from '@dnd-kit/utilities';
-import {DateInput} from "./DatePicker";
+import { Task } from "../../types";
+import { DateInput } from "./DatePicker";
 import { useEffect, useState } from "react";
 
 export function TodoTaskLabel({task}:{task:Task}) {
@@ -15,19 +13,7 @@ export function TodoTaskLabel({task}:{task:Task}) {
     }
   }, [date]);
 
-  const {attributes, listeners, setNodeRef, transform} = useDraggable({
-    id: task.id,
-  });
-
-  const style = {
-    transform: CSS.Translate.toString(transform)
-  };
-
   return (
-    <article id={task.id} data-category={task.category as EMCategory} className="relative rounded bg-white font-bold my-2 p-2 text-sm text-habitxt" ref={setNodeRef} style={style} {...listeners} {...attributes}>
-      <h3>{ task.name }</h3>
-      <span className="inline-block text-xs right-2 absolute">{ task.type }</span>
-      <DateInput date={date} setDate={setDate} />
-    </article>
+    <DateInput date={date} setDate={setDate} />
   )
 }
